Implement playground examples
This is a pretty big commit, but it's not like I've implemented everything in a day. I've been tinkering with this for a while, but I didn't want to push anything until I had something ready (in the meantime, I just use the `wip` branch if I want to save my work).